app/backend/requirements.txt
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
et_xmlfile==2.0.0
fastapi==0.115.6
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
loguru==0.7.3
openpyxl==3.1.5
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyserial==3.5
python-dotenv==1.2.1
python-multipart==0.0.12
PyYAML==6.0.3
SQLAlchemy==2.0.46
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.1
uvloop==0.22.1; platform_system != "Windows"
watchfiles==1.1.1
websockets==16.0
